pyproject.toml
[project]
name = "pokedex-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "async-lru>=2.0.5",
    "pydantic-ai>=0.3.4",
]
